pyproject.toml
[tool.poetry]
name = "tursi"
version = "0.3.0-alpha.3"
description = "A simple framework to deploy AI models locally with one command, no containers needed"
authors = ["Kresna Sucandra <kresnasucandra@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/BlueTursi/tursi-ai"
classifiers = [
    "License :: OSI Approved :: MIT License",
]
packages = [{include = "tursi"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
flask = "^3.0.0"
transformers = "^4.38.0"
torch = "^2.2.1"
python-dotenv = "^1.0.0"
flask-limiter = "^3.5.0"
optimum = "^1.17.0"
accelerate = "^0.27.0"
onnxruntime = "^1.17.0"
onnx = "^1.15.0"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.0.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
black = "^24.0.0"
build = "^1.0.3"
twine = "^4.0.2"
mypy = "^1.9.0"
types-requests = "^2.31.0"
types-flask = "^1.1.0"
types-psutil = "^5.9.0"

[tool.poetry.scripts]
tursi = "tursi.cli:main"
tursi-test = "tursi.test:main"
tursid = "tursi.daemon:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=tursi --cov-report=term-missing"
ini_file = "tests/pytest.ini"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["transformers.*", "optimum.*", "torch.*"]
ignore_missing_imports = true
